Fixed bugs that stopped it doing anything.
Color is now in the English spelling. "colour". Removed bad call to
timeout functions. Added naturalHeight and naturalWidth. Events are now
not removed if there is a file error, and can now set the src to another
location if the gif did not find or was unable to complete the file
request. Also fixed some help